requirements.txt
streamlit>=1.32.0
pydantic>=2.0.0
pyyaml>=6.0
huggingface-hub>=0.22.0
transformers>=4.40.0
torch>=2.2.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
plotly>=5.20.0
pytest>=8.0.0
